Fix overflow in av1_rc_clamp_pframe_target_size()
Port the changes in
https://aomedia-review.googlesource.com/c/aom/+/154784 to
av1_rc_clamp_pframe_target_size().

Port some of the changes in
https://aomedia-review.googlesource.com/c/aom/+/107521 to
av1_calc_pframe_target_size_one_pass_cbr().

The goal is to make sure that all four combinations of
av1_calc_{iframe,pframe}_target_size_one_pass_{vbr,cbr}() are fixed in
the same way.

+++ av1/encoder/ratectrl.h
@@ -694,7 +694,7 @@ int av1_rc_bits_per_mb(const struct AV1_COMP *cpi, FRAME_TYPE frame_type,
 int av1_rc_clamp_iframe_target_size(const struct AV1_COMP *const cpi,
                                     int64_t target);
 int av1_rc_clamp_pframe_target_size(const struct AV1_COMP *const cpi,
-                                    int target, uint8_t frame_update_type);
+                                    int64_t target, uint8_t frame_update_type);
 
 // Find q_index corresponding to desired_q, within [best_qindex, worst_qindex].
 // To be precise, 'q_index' is the smallest integer, for which the corresponding
